Updated all .java

+++ src/main/java/com/example/Animal/Website/CRUD/API/Bird.java
@@ -1,27 +1,28 @@
 package com.example.Animal.Website.CRUD.API;
 
 import jakarta.persistence.Entity;
-import jakarta.persistence.Id;
 import jakarta.persistence.GeneratedValue;
 import jakarta.persistence.GenerationType;
+import jakarta.persistence.Id;
 import jakarta.persistence.Table;
 
 @Entity
 @Table(name = "birds") 
 public class Bird {
 
-@Id
-@GeneratedValue(strategy = GenerationType.IDENTITY)
+    @Id
+    @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long birdId;
+    
     private String name;
     private String description;
     private String breed;
-    private double age;
+    private int age;
 
     public Bird() {
     }
 
-    public Bird(Long birdId, String name, String description, String breed, double age) {
+    public Bird(Long birdId, String name, String description, String breed, int age) {
         this.birdId = birdId;
         this.name = name;
         this.description = description;
@@ -29,7 +30,7 @@ public class Bird {
         this.age = age;
     }
 
-    public Bird(String name, String description, String breed, double age) {
+    public Bird(String name, String description, String breed, int age) {
         this.name = name;
         this.description = description;
         this.breed = breed;
@@ -72,7 +73,7 @@ public class Bird {
         return age;
     }
 
-    public void setAge(double age) {
+    public void setAge(int age) {
         this.age = age;
     }
 }
